system-roms-mess_artwork: Fix failures when a mess artwork directory doesn't exist

+++ bin/setup/system-roms-mess_artwork.sh
@@ -119,6 +119,10 @@ __build_background_view() {
 }
 
 restore() {
+  if [ ! -d "$system_artwork_dir" ]; then
+    return
+  fi
+
   find "$system_artwork_dir" -mindepth 1 -type d -exec rm -rfv '{}' +
 }
